working on splitting things into smaller files

+++ include/vm.h
@@ -31,4 +31,4 @@ dlvm_t *dlvm_init(uint8_t *p, uint64_t ps, uint64_t ss);
 
 void dlvm_push(dlvm_t *vm, ttype_t *o);
 ttype_t *dlvm_pop(dlvm_t *vm);
-uint64_t dlvm_next_op(dlvm_t *vm);
+uint8_t dlvm_next_op(dlvm_t *vm);
